Split some syscalls
