+++ kernel/domain/descriptor.c
@@ -147,143 +147,3 @@ int get_process(process_t **process, int process_fd) {
     *process = (process_t *)object;
     return 0;
 }
-
-int dup(int process_fd, int src, int dest) {
-    process_t *process;
-
-    int status = get_process(&process, process_fd);
-
-    if(status < 0) {
-        return status;
-    }
-
-    object_header_t *object;
-    object_ref_t *src_ref;
-    status = dereference_object_descriptor(&object, &src_ref, get_current_process(), src);
-
-    if(status < 0) {
-        return status;
-    }
-
-    if(object_ref_is_owner(src_ref)) {
-        return -JINUE_EBADF;
-    }
-
-    object_ref_t *dest_ref;
-    status = dereference_unused_descriptor(&dest_ref, process, dest);
-
-    if(status < 0) {
-        return status;
-    }
-
-    dest_ref->object = src_ref->object;
-    dest_ref->flags  = src_ref->flags;
-    dest_ref->cookie = src_ref->cookie;
-
-    object_open(object, dest_ref);
-
-    return 0;
-}
-
-static int check_mint_permissions(const object_header_t *object, int perms) {
-    if((perms & ~object->type->all_permissions) != 0) {
-        return -JINUE_EINVAL;
-    }
-
-    /* TODO remove this once permissions are defined for process objects */
-    if(object->type == object_type_process) {
-        return 0;
-    }
-
-    if(perms == 0) {
-        return -JINUE_EINVAL;
-    }
-
-    return 0;
-}
-
-int mint(int owner, const jinue_mint_args_t *mint_args) {
-    object_header_t *object;
-    object_ref_t    *src_ref;
-    
-    int status = dereference_object_descriptor(&object, &src_ref, get_current_process(), owner);
-
-    if(status < 0) {
-        return status;
-    }
-
-    status = check_mint_permissions(object, mint_args->perms);
-
-    if(status < 0) {
-        return status;
-    }
-
-    if(!object_ref_is_owner(src_ref)) {
-        return -JINUE_EPERM;
-    }
-
-    process_t *process;
-
-    status = get_process(&process, mint_args->process);
-
-    if(status < 0) {
-        return status;
-    }
-
-    object_ref_t *dest_ref;
-    
-    status = dereference_unused_descriptor(&dest_ref, process, mint_args->fd);
-
-    if(status < 0) {
-        return status;
-    }
-
-    dest_ref->object = src_ref->object;
-    dest_ref->flags  =
-          mint_args->perms
-        | (src_ref->flags & OBJECT_FLAG_DESTROYED)
-        | OBJECT_REF_FLAG_IN_USE;
-    dest_ref->cookie = mint_args->cookie;
-
-    object_open(object, dest_ref);
-
-    return 0;
-}
-
-int close(int fd) {
-    object_header_t *object;
-    object_ref_t *ref;
-    int status = dereference_object_descriptor(&object, &ref, get_current_process(), fd);
-
-    if(status < 0) {
-        return status;
-    }
-    
-    object_close(object, ref);
-
-    ref->flags = OBJECT_REF_FLAG_NONE;
-
-    return 0;
-}
-
-int destroy(int fd) {
-    object_header_t *object;
-    object_ref_t *ref;
-    int status = dereference_object_descriptor(&object, &ref, get_current_process(), fd);
-
-    if(status < 0) {
-        return status;
-    }
-
-    if(!object_ref_is_owner(ref)) {
-        return -JINUE_EPERM;
-    }
-
-    object_close(object, ref);
-
-    object_mark_destroyed(object);
-
-    ref->flags = OBJECT_REF_FLAG_NONE;
-
-    return 0;
-}
